break out IdentityDetailSerializer
