+++ api/serializers.py
@@ -23,18 +23,6 @@ from rest_framework import pagination
 from threepio import logger
 
 
-# Serializers
-class IdentityDetailSerializer(serializers.ModelSerializer):
-    created_by = serializers.CharField(source='creator_name')
-    quota = serializers.Field(source='get_quota_dict')
-    provider_id = serializers.Field(source='provider.uuid')
-    id = serializers.Field(source="uuid")
-
-    class Meta:
-        model = Identity
-        exclude = ('credentials', 'created_by', 'provider', 'uuid')
-
-
 class AtmoUserSerializer(serializers.ModelSerializer):
     selected_identity = IdentityRelatedField(source='select_identity')
